changes story points to be a decimal, adds a helper for default tags

+++ db/migrate/20120110150916_add_story_points_to_tickets.rb
@@ -1,5 +1,5 @@
 class AddStoryPointsToTickets < ActiveRecord::Migration
   def change
-    add_column :tickets, :story_points, :integer
+    add_column :tickets, :story_points, :decimal
   end
 end
